refactor(trace): migrate etwtracer component to v2
Restructures comp/trace/etwtracer from the V1 layout (component.go +
etwtracerimpl/) to the V2 layout (def/component.go, impl/, fx/).
