+++ comp/trace/etwtracer/def/component.go
