+++ comp/trace/etwtracer/fx/fx.go
@@ -1,0 +1,21 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2023-present Datadog, Inc.
+
+// Package fx provides the fx module for the etwtracer component.
+package fx
+
+import (
+	"go.uber.org/fx"
+
+	etwtracerimpl "github.com/DataDog/datadog-agent/comp/trace/etwtracer/impl"
+	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
+)
+
+// Module defines the fx options for this component.
+func Module() fxutil.Module {
+	return fxutil.Component(
+		fx.Provide(etwtracerimpl.NewComponent),
+	)
+}
